Move tests for create_transaction_event_for_transaction_session to separate file
